Cropping: fix composite image creation
Due to (apparently) a bug in GraphicsMagic, one is not able to create a
composite image based on a blank image that was created in memory if the
composite parts are in certain formats. This commit works around this by
creating a canvas image that is built from the first composite part and
therefore has the exact same storage properties.