Change segment/slice foreign keys to bigint
South failed to change foreign keys to the segment and slice tables from
varchar to bigint during the previous migration. Change these keys,
remove varchar-based indices, add constraints, and add indices.

Note that although value conversion is used, because many of the slice
hashes use the MSB, value conversion from string representation of the
hash to bigint will fail due to signed 64-bit limits. There is no
straightforward way to handle this case in Postgres, and because all the
old data is deprecated anyway all tables should be empty when running
this migration.