Djsopnet: return null ID for missing objects
Do not return -1, which is a valid ID (though not in the default
sequence) for many models. Instead return null. In general the API
should indicate an object was not found with a status code, but this is
a temporary improvement to prevent attempting to fetch non-existent
objects from clients.