Catsop widget: load segments at location
Add a view to retrieve slices at a stack location. Add a button to the
widget that uses these slices to retrieve a segments for the segment
graph at the current stack location.

This could easily be reduced to one request/query, but this more general
approach is sufficiently functional and fast for current requirements.