Djsopnet: support history of solution sets
Support storing core solutions as sets, identified in the
djsopnet_solution table. Segments are now associated with solutions,
rather than directly with cores. Solutions for the same core take
precedence over each other through the djsopnet_solutionprecedence
table. Only one solution can have precedence for a given core at a time,
enforced through a uniqueness constraint. Attempting to INSERT a
precedent solution when one already exists results instead in an UPDATE
to that core's precedence via a RULE.

Also create a correction table that annotates which explicit segment
solutions are being corrected by user added constraints.

As with migrations 0016 and 0017, these changes must be implemented in
two migrations because a column is added to populated table, updated to
reflect a relational graph change, and can only be set as non-nullable
in a subsequent transaction. Note that the backwards migrations delete
all solution sets that do not have precedence.