Fix djsopnet and CATSOP widget for block scale
Djsopnet now correctly accepts location and bounding box requests in
stack space, converting to scaled block space internally. Segmentation
objects are still served in scale coordinates to the front-end, so the
CATSOP widget now requests BlockInfo for the current stack so that it
can translate appropriately. The CATSOP layer adjusts for block scale
at the same time it adjust for stack scale when drawing slices.