Djsopnet: remove extra ID column for 1-to-1 tables
Remove the extra django-generated primary key ID columns for tables
unique indentified by a one-to-one foreign key. These tables are
segmentfeatures (for segment) and solutionprecedence (for core). Instead
make the one-to-one foreign key the primary key and remove the redundant
unique constraint on that foreign key. Reduces row size and improves
performance.