Djsopnet: make stack-specific segmentation schemas
CATSOP segmentations for a project are now configured via
SegmentationConfiguration objects. ProjectStacks participate in
segmentations via SegmentationStack objects. Segmentation data for each
of these SegmentationStacks exists in a separate schema of the form
segstack_{id}, which inherits its DDL from the segstack_template schema.
This allows segmentation data to not collide without needless and
expensive duplication of some segmentation-project key in every table
(see catsop/CATMAID#3 for detailed discussion).

Because Django ORM can not understand schema duplication in a
concurrency-safe way (or any way that is not awkward and obtuse), and
because Django ORM imposes restrictions on schemas, like the lack of
multicolumn primary keys, that cause waste for large segmentation
databases, segmentation tables are no longer managed by Django ORM or
represented as Django models. Instead, these schemas/tables should be
accessed through SQL directly. DDL for the segmentation schema is now in
segmentation_schema.sql, which should be kept up-to-date with the actual
schema structure. Changes to the schema will still be handled in South
migrations, but with plain SQL instead of South's helpers. The script
for copying this schema to a segmentation stack specific schema is in
instantiate_schema.sql. Note that foreign keys must be explicitly
redeclared in this script. A post-save signal receiver on
SegmentationStack automatically executes this script when a new
SegmentationStack is created via Django ORM.

Note that testing scripts are updated to reflect these changes, which is
sufficiently to allow pysopnet to run. However, views and other scripts
have not yet been updated, so the Djsopnet UI is broken as of this
commit. It will be fixed in the following commits.

Closes catsop/CATMAID#3.
Closes catsop/catsop#37.
Closes catsop/catsop#33.