Skeleton intersection: update for schema changes
Update untested skeleton intersection code from the 2014 hackathon to
reflect cumulative changes to the schema since then, and to reuse
functionality where appropriate. Constraint generation from skeletons
now runs but is untested. This intersection code also does yet not
account for a known case where multiple skeletons or arbors in the same
skeleton are compatible with a common segment, which should be excluded.