Skeleton intersection: remove section lookup
Remove code originally intended to further group supernodes which lay in
the same section. Note that this code was not affecting the result.
Justification from an email exchange with @unidesigner pasted below for
future reference:

1 *
   \
2   *-*   *-*
       \ /
3       *

So the idea of the section lookup was to somehow combine the constraints
for all four nodes in z=2 instead of just into two constraints for the
two super nodes. Let's call the four nodes in section 2 A, B, C and D.
Let's say there's a slice x that includes A & B but not C & D. Likewise
there's a slice y for C & D, and a slice z that includes all 4. Then the
super node AB will have sliceset {x, z} and the super node CD will have
sliceset {y, z}. If we make two constraints for the cartesian product of
segments with the sliceset for the node in section 3, either x & y will
be chosen by solver (via a branch segment if one exists) or z will.
Since there's no prior reason to believe z is the correct slice (x and y
could be different intersections of the process with the plane of
section 2 after bending through section 3), this seems fine and leaves
it up to the solver to decide based on the value of x, y, z, existence
of the relevant x-, y-branch and z-continuation segments, and other
context.