Skeleton intersection: handle skeleton conflicts
When more than one skeleton is compatible with a segment, prevent that
segment from being chosen by the solver. See code comments for details
of how this is achieved with constraint coefficients. This is a brute
force implementation that counts the number of skeletons constraining a
segment whenever a new skeleton constraint is created for that segment.