Djsopnet: rename test settings
Rename settings only used for test scripts and test data bootstrapping
from 'SOPNET_{NAME}' to 'SOPNET_TEST_{NAME}'. This makes it much more
clear which settings are required for CATSOP in general and which are
specific to test data.

Make a SOPNET_TEST_PARALLEL_JOBS setting which is a dictionary
specifying the number of threads to use for each test script, or False
for sequential testing.